Move delete modal confirmation state to common
