+++ src/common/store/ui/delete-playlist-confirmation-modal/reducers.ts
