+++ src/common/store/ui/delete-playlist-confirmation-modal/selectors.ts
@@ -1,0 +1,6 @@
+import { CommonState } from 'common/store/'
+
+export const getIsOpen = (state: CommonState) =>
+  state.ui.deletePlaylistConfirmationModal.isOpen
+export const getPlaylistId = (state: CommonState) =>
+  state.ui.deletePlaylistConfirmationModal.playlistId
