--- src/containers/delete-playlist-confirmation-modal/store/selectors.ts
@@ -1,6 +1,0 @@
-import { AppState } from 'store/types'
-
-export const getIsOpen = (state: AppState) =>
-  state.application.pages.deletePlaylistConfirmation.isOpen
-export const getPlaylistId = (state: AppState) =>
-  state.application.pages.deletePlaylistConfirmation.playlistId
